bump-cask-pr: handle nil cask version, add tests
The `shortened_version` method in the `bump-cask-pr` command can
produce an error if `cask.version` is `nil`. For example, this can
happen on Linux when a cask only sets the `version` in macOS
on_system blocks. This modifies the method to explicitly raise a
`CaskInvalidError` when `cask.version` is `nil`, as we need to know
the existing version to be able to determine how to handle the new
`version`. [It's possible to fall back to the full version instead
but that can allow certain types of duplicate PRs when `cask.version`
is or isn't `nil` on a given host.]

This also adds tests for the `shortened_version` method, bringing
it to 100% coverage.