docs(cask): clarify that 3rd-party pre-compiled binaries should be casks
Goreleaser recently changed their default brew formula generation type
from formula -> cask. This may be confusing for people who vaguely think
that formula == binary, cask == macOS Application.

I'm not sure where to document this in more detail, but my understanding
is that if you `brew install` and it uses a pre-built binary instead of
  building from source:
- If it's from homebrew/core -> formula
- If it's from a third-party tap -> cask

so basically bottles for third-party taps aren't a thing.

Refs: https://github.com/orgs/goreleaser/discussions/5563#discussioncomment-13368387