DownloadStrategyDetector: handle tangled.sh Git URLs
We have one formula (`lsr`) that uses a tangled.sh URL but they don't
support a .git extension, so it isn't detected as a Git URL by
`DownloadStrategyDetector`. We can use `using: :git` with the `stable`
URL but livecheck can't use the `Git` strategy for tangled.sh URLs
because its `match` method depends on `DownloadStrategyDetector`
detecting the URL as a `GitDownloadStrategy`.

This adds support for tangled.sh URLs to `DownloadStrategyDetector`
to address this issue. This will also allow us to remove/omit the
`using: :git` argument from related URLs.